Introduce `checkClientGet` for 404 handling
This new helper properly and uniformly handles Sentry API 404 errors
with respect to the Terraform SDK. Prior to this commit, in most places
an API error was always interpretted as a missing resource. Now, a 404
is considered a missing resource and other API errors are surfaced as
real errors. This will ensure that, for example, a user with a malformed
API token (bad permissions, etc) doesn't wipe out Terraform state.

Fixes the following Github issues:

- https://github.com/jianyuan/terraform-provider-sentry/issues/40
- https://github.com/jianyuan/terraform-provider-sentry/issues/57