fix(resource/sentry_issue_alert): `0` values in `conditions_v2`
There is [arguably a bug in the Sentry API](https://github.com/getsentry/sentry/issues/89042) where integer values of 0 will be stripped from submitted payload objects because of Python's implicit bool conversions. These values can successfully be submitted as strings. When these objects are retrieved from the API, the values are integers.

To model this, we have two different schemas in the OpenAPI spec yaml for rule conditions based based on whether we're sending them to the API or retriving them from the API:
* `ProjectRuleCondition`, which is a set of condition schemas where `value`s are `integers`
* `ProjectRuleConditionToApi` which is a set of condition schemas where `value`s are strings

The generated types from these are used in the alert rule resource code in the `Fill` and `ToApi` methods as necessary.

There may be a way to do this without so much duplication, I am definitely open to suggestions!